ecodata: indicator data for reporting ecosystem health

ecodata is developed by the Ecosystems Dynamics and Assessment Branch of the Northeast Fisheries Science Center for use in State of the Ecosystem (SOE) reporting. SOE reports are high-level overviews of ecosystem indicator status and trends occurring on the Northeast Continental Shelf. Unless otherwise stated, data are representative of specific Ecological Production Units (EPUs), referring to the Mid-Atlantic Bight (MAB), Georges Bank (GB), Gulf of Maine (GOM), and Scotian Shelf (SS). SOE reports are developed for US Fishery Management Councils (FMCs), and therefore indicator data for Scotian Shelf are included when available, but this is not always the case.

Details

To create a centralized repository for indicator data

To facilitate creation of annual reports

To learn more about using ecodata, start with the vignette: browseVignettes(package="ecodata") or click the index link below

See also

Author

Maintainer: Brandon Beltz brandon.beltz@noaa.gov

Authors:

Andy Beet andrew.beet@noaa.gov (ORCID)

Kimberly Bastille kimberly.bastille@noaa.gov

Sean Hardison