Harbor porpoise density estimates

Data include seasonal density estimates for harbor porpoise on the Northeast US Continental Shelf.

hp_density

Format

6244 rows and 5 columns

Latitude: Decimal latitude.

Longitude: Decimal longitude.

Value: Density (Animals/km^2)

Bin: Density bins for plotting.

Season: Season; fall or spring.

Source

Data were sourced from https://www.nefsc.noaa.gov/AMAPPSviewer/ on 12/14/2018.

Details

From the AMAPPS Marine Mammal Model Viewer Site: "Distribution and abundance of wildlife is largely driven by physical and biological environmental factors, including climate, habitat characteristics and prey distribution (Ainley et al. 2005). To account for this, these spatially- and temporally-explicit density maps were based on animal density - environmental generalized additive statistical models that were fit to visual shipboard and aerial survey line-transect data, associated survey conditions, animal group characteristics, spatially- and temporally-explicit static and dynamic environmental data, and species-specific availability bias correction factors. For a more completed description see Palka et al. 2017."

References

Ainley D.G., L. Spear, C.T. Tynan, J.A. Barth, S.D. Pierce, R. Glenn Ford, T. Cowles. 2005. Physical and biological variables affecting seabird distributions during the upwelling season of the northern California Current. Deep Sea Research II 52(2):123–143.

Palka, D.L., S. Chavez-Rosales, E. Josephson, D. Cholewiak, H.L. Haas, L. Garrison, M. Jones, D. Sigourney, G. Waring (retired), M. Jech, E. Broughton, M. Soldevilla, G. Davis, A. DeAngelis, C.R. Sasso, M.V.Winton, R.J. Smolowitz, G. Fay, E. LaBrecque, J.B. Leiness, Dettloff, M. Warden, K. Murray, and C. Orphanides. 2017. Atlantic Marine Assessment Program for Protected Species: 2010-2014. US Dept. of the Interior, Bureau of Ocean Energy Management, Atlantic OCS Region, Washington, DC. OCS Study BOEM 2017-071. 211 pp