Lepanthes rupestris

A 10-year presence/absence dataset (1999-2008) of Lepanthes rupestris at 975 phorophyte patches in the Luquillo Experimental Forest, Puerto Rico.

lepanthes_rupestris

Format

A data frame with 975 rows and 23 variables:

UID

Unique id for phorophyte

x.coord

X-coordinate

y.coord

Y-coordinate

z.coord

Z-coordinate

Phorophyte

Phorophyte type either rock or tree

X70199

Sampling in 1999

X20100

First sampling of 2000

X80100

Second sampling of 2000

X10101

First sampling of 2001

X60101

Second sampling of 2001

X10102

First sampling of 2002

X60102

Second sampling of 2002

X120202

Third sampling of 2002 (late in the year)

X61003

First sampling of 2003

X12404

First sampling of 2004

X71704

Second sampling of 2004

X12205

First sampling of 2005

X91705

Second sampling of 2005

X72406

First sampling of 2006

X20307

First sampling of 2007

X61507

Second sampling of 2007

X10208

First sampling of 2008

Total_Moss_area

Total are covered by moss in the phorophyte

Source

https://doi.org/10.5061/dryad.9p8cz8wc6

Citation

Acevedo, Miguel; Beaudrot, Lydia; Melendez-Ackerman, Elvia; Tremblay, Raymond (2020), Local extinction risk under climate change in a neotropical asymmetrically dispersed epiphyte, Dryad, Dataset, https://doi.org/10.5061/dryad.9p8cz8wc6