New England inshore state survey data

These data contain time series of inshore (within 12 miles) fishery-independent survey data collected by the Maine Department of Marine Resources and New Hampshire Fish & Game Department. Time series are grouped by State of the Ecosystem feeding guild. Species assignments within groupings are available through inshore_survdat_species.

Usage

ne_inshore_survey

Format

A data set containing 264 rows and 6 columns.

Var: Includes SOE species grouping, season that survey occurred (fall or spring), and variable calculated. The specific variables included are stratified mean biomass with associated confidence intervals, coefficients of variation, and standard errors.

Time: Year that survey occurred.

Value: Value of variable Var.

Units: Units of variable Var.

Source: Survey Source.

EPU: Ecological Production Unit (EPU) where sampling occurred. Here NE refers to New England.

Details

Metrics for biomass were calculated by:

All species catch weights were summed up for each tow for each group.

Then the average weight per tow and associated variances and standard deviation for each survey, region, stratum, and species group was calculated.

The average weight for each group was then weighted by total area surveyed in each region and stratum. (average weight * area surveyed).

The new weighted averages were then summed up by survey and species group then the totals were divided by the total area of the survey (11699.831 km2) to provide the stratified mean biomass for each species group in each survey. The coefficient of variation, standard error, and 95% confidence intervals were also calculated for each species group and provided.

Read more about inshore surveys at https://wildlife.state.nh.us/marine/research.html and https://www.maine.gov/dmr/science-research/projects/trawlsurvey/index.html